aaaa

+++ app/src/main/java/com/example/cfmatch/entities/User.java
@@ -4,15 +4,13 @@ import java.io.Serializable;
 
 public class User implements Serializable {
 
-    public int id;
+    public long id;
     public String name;
     public String email;
     public String password;
     public String description;
 
-    public User () {
-
-    }
+    public User () {}
 
     public User (String email, String password, String name, String description) {
         this.email = email;
